third_party/WebKit/LayoutTests/external/wpt/cssom-view/scrollingElement-expected.txt

This is a testharness.js-based test.
FAIL scrollingElement in quirks mode assert_equals: expected Element node <body style="overflow: scroll;"></body> but got null
PASS scrollingElement in no-quirks mode 
Harness: the test ran to completion.
